fixing the build and added shared unix layer
