+++ src/conditionals/os/unix/macos/impl/platform.c
@@ -13,49 +13,8 @@
  * License for more details.                                                *
  ****************************************************************************/
 
-#include "clx_os/path.h"
+#include "clx_os/platform.h"
 
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-
-char * clx_os_path_canonicalize(char const * path) {
-    return realpath(path, NULL);
-}
-
-bool clx_os_path_is_absolute(char const * path) {
-    return path && path[0] == '/';
-}
-
-char const * clx_os_path_find_last_separator(char const * path) {
-    if (!path) {
-        return NULL;
-    }
-
-    return strrchr(path, '/');
-}
-
-char clx_os_path_separator(void) {
-    return '/';
-}
-
-char * clx_os_path_executable_dir(void) {
-    char buf[4096];
-    ssize_t len = readlink("/proc/self/exe", buf, sizeof(buf) - 1);
-    if (len <= 0) {
-        return NULL;
-    }
-    buf[len] = '\0';
-    char * last = strrchr(buf, '/');
-    if (!last) {
-        return strdup(".");
-    }
-    size_t dirLen = (size_t)(last - buf);
-    char * dir = malloc(dirLen + 1);
-    if (!dir) {
-        return NULL;
-    }
-    memcpy(dir, buf, dirLen);
-    dir[dirLen] = '\0';
-    return dir;
+char const * clx_os_platform_name(void) {
+    return "macos";
 }
